[startback-jobs] Handle job errors.
The commit introduces an (optional) hasFailed flag on jobs, that is set
to true for jobs that completes with a failure (isReady is set to true).

When serving job results for jobs that have failed, a special 2xx status
code (272) is used: the web service call succeeds (it must be a 2xx) and
the result dumped in json. The client can distinguish successes from
failures by checking the status code.